116-FEAT:  package updates, image upload, number input
- got image upload working
- added in the new number input
- split up the image upload dialog
- remove old code
- package updates
