+++ apps/website/package.json
@@ -18,7 +18,7 @@
     "db:stop": "supabase stop"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.4.0",
+    "@hookform/resolvers": "^3.4.2",
     "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-collapsible": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
@@ -34,65 +34,65 @@
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-toggle-group": "1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@react-email/components": "0.0.17",
+    "@react-email/components": "0.0.19",
     "@supabase/ssr": "^0.3.0",
-    "@supabase/supabase-js": "^2.43.2",
+    "@supabase/supabase-js": "^2.43.4",
     "@t3-oss/env-nextjs": "^0.10.1",
-    "@tanstack/query-core": "^5.36.1",
-    "@tanstack/react-query": "^5.36.2",
-    "@tanstack/react-query-devtools": "^5.36.2",
-    "@typescript-eslint/eslint-plugin": "^7.9.0",
-    "@typescript-eslint/parser": "^7.9.0",
-    "@uploadthing/react": "^6.5.4",
-    "@vercel/analytics": "^1.2.2",
-    "@vercel/speed-insights": "^1.0.10",
+    "@tanstack/query-core": "^5.40.0",
+    "@tanstack/react-query": "^5.40.0",
+    "@tanstack/react-query-devtools": "^5.40.0",
+    "@typescript-eslint/eslint-plugin": "^7.11.0",
+    "@typescript-eslint/parser": "^7.11.0",
+    "@uploadthing/react": "^6.6.0",
+    "@vercel/analytics": "^1.3.1",
+    "@vercel/speed-insights": "^1.0.11",
     "autoprefixer": "^10.4.19",
     "cheerio": "1.0.0-rc.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "dotenv": "^16.4.5",
-    "embla-carousel": "8.0.4",
-    "embla-carousel-autoplay": "8.0.4",
-    "embla-carousel-react": "8.0.4",
+    "embla-carousel": "8.1.3",
+    "embla-carousel-autoplay": "8.1.3",
+    "embla-carousel-react": "8.1.3",
     "eslint-plugin-drizzle": "^0.2.3",
     "fraction.js": "^4.3.7",
-    "framer-motion": "^11.2.2",
+    "framer-motion": "^11.2.10",
     "fuse.js": "^7.0.0",
     "fuzzball": "^2.1.2",
     "input-otp": "^1.2.4",
-    "jotai": "^2.8.0",
+    "jotai": "^2.8.2",
     "jotai-tanstack-query": "^0.8.5",
-    "lucide-react": "^0.378.0",
+    "lucide-react": "^0.381.0",
     "next": "^14.2.3",
-    "next-axiom": "^1.2.0",
+    "next-axiom": "^1.3.0",
     "next-safe-action": "^6.2.0",
     "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
     "react-dom": "^18.3.1",
-    "react-email": "^2.1.2",
-    "react-hook-form": "^7.51.4",
+    "react-email": "^2.1.4",
+    "react-hook-form": "^7.51.5",
     "react-hot-toast": "^2.4.1",
     "react-select": "^5.8.0",
     "resend": "^3.2.0",
-    "sharp": "^0.33.3",
+    "sharp": "^0.33.4",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
     "@mixie/email": "workspace:*",
     "unsplash-js": "^7.0.19",
-    "uploadthing": "^6.10.4",
+    "uploadthing": "^6.12.0",
     "uuid": "^9.0.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@hookform/devtools": "^4.3.1",
     "@tanstack/eslint-plugin-query": "^5.35.6",
-    "@types/node": "^20.12.12",
-    "@types/react": "^18.3.2",
+    "@types/node": "^20.13.0",
+    "@types/react": "^18.3.3",
     "@types/react-beautiful-dnd": "^13.1.8",
     "@types/react-dom": "^18.3.0",
     "postcss": "^8.4.38",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.0",
     "tailwindcss": "^3.4.3",
     "@mixie/tsconfig": "workspace:*",
     "typescript": "^5.4.5"
