+++ apps/website/src/components/recipe-form/components/image-upload/header-controls.tsx
@@ -19,7 +19,7 @@ export const HeaderControls: FC<HeaderControlsProps> = ({
       {currentStepId != null && (
         <Button
           variant="ghost"
-          className="absolute left-4 top-4 h-fit w-fit rounded-sm p-0 opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground"
+          className="absolute top-4 left-4 h-fit w-fit rounded-sm p-0 opacity-70 ring-offset-background transition-opacity disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2"
           aria-label="Return to previous step"
           onClick={goToPreviousStep}
         >
@@ -29,7 +29,7 @@ export const HeaderControls: FC<HeaderControlsProps> = ({
       )}
 
       {/* Close button */}
-      <DialogClose className="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground">
+      <DialogClose className="absolute top-4 right-4 rounded-sm opacity-70 ring-offset-background transition-opacity disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2">
         <X className="h-6 w-6" />
         <span className="sr-only">Close</span>
       </DialogClose>
