+++ apps/website/src/components/ui/number-input.tsx
@@ -2,37 +2,47 @@ import React from "react";
 import { Button } from "./button";
 import { MinusIcon, PlusIcon } from "lucide-react";
 
-const NumberInput = () => {
-  const [goal, setGoal] = React.useState(350);
+interface NumberInputProps {
+  value: number;
+  onMinusClick: () => void;
+  onPlusClick: () => void;
+  minusDisabled?: boolean;
+  plusDisabled?: boolean;
+}
 
-  function onClick(adjustment: number) {
-    setGoal(Math.max(200, Math.min(400, goal + adjustment)));
-  }
-
+const NumberInput = ({
+  value,
+  onMinusClick,
+  onPlusClick,
+  minusDisabled = false,
+  plusDisabled = false,
+}: NumberInputProps) => {
   return (
     <div className="flex items-center justify-center space-x-2">
       <Button
         variant="outline"
         size="icon"
+        type="button"
         className="h-8 w-8 shrink-0 rounded-full"
-        onClick={() => onClick(-10)}
-        disabled={goal <= 200}
+        onClick={onMinusClick}
+        disabled={minusDisabled}
       >
         <MinusIcon className="h-4 w-4" />
         <span className="sr-only">Decrease</span>
       </Button>
       <div className="flex-1 text-center">
-        <div className="text-5xl font-bold tracking-tighter">{goal}</div>
-        <div className="text-[0.70rem] uppercase text-muted-foreground">
+        <div className="font-bold text-5xl tracking-tighter">{value}</div>
+        {/* <div className="text-[0.70rem] text-muted-foreground uppercase">
           Calories/day
-        </div>
+        </div> */}
       </div>
       <Button
         variant="outline"
         size="icon"
+        type="button"
         className="h-8 w-8 shrink-0 rounded-full"
-        onClick={() => onClick(10)}
-        disabled={goal >= 400}
+        onClick={onPlusClick}
+        disabled={plusDisabled}
       >
         <PlusIcon className="h-4 w-4" />
         <span className="sr-only">Increase</span>
