+++ package.json
@@ -15,12 +15,10 @@
     "@manypkg/cli": "^0.21.4",
     "@turbo/gen": "^1.13.3",
     "dotenv-cli": "^7.4.2",
-    "eslint": "^9.2.0",
-    "prettier": "^3.2.5",
     "turbo": "^1.13.3"
   },
   "devDependencies": {
-    "prettier-plugin-tailwindcss": "^0.5.14",
+    "prettier-plugin-tailwindcss": "^0.6.1",
     "@biomejs/biome": "1.7.3"
   },
   "trustedDependencies": [
